+++ app/src/main/java/org/breezyweather/sources/openmeteo/OpenMeteoWeatherApi.kt
@@ -31,6 +31,7 @@ interface OpenMeteoWeatherApi {
         @Query("longitude") longitude: Double,
         @Query("daily") daily: String,
         @Query("hourly") hourly: String,
+        @Query("minutely_15") minutely_15: String,
         @Query("forecast_days") forecastDays: Int,
         @Query("past_days") pastDays: Int,
         @Query("current_weather") currentWeather: Boolean,