+++ app/src/main/java/org/breezyweather/sources/openmeteo/json/OpenMeteoWeatherMinutely.kt
@@ -16,16 +16,11 @@
 
 package org.breezyweather.sources.openmeteo.json
 
-import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
-/**
- * Open-Meteo weather
- */
 @Serializable
-data class OpenMeteoWeatherResult(
-    @SerialName("current_weather") val currentWeather: OpenMeteoWeatherCurrent? = null,
-    val daily: OpenMeteoWeatherDaily? = null,
-    val hourly: OpenMeteoWeatherHourly? = null,
-    @SerialName("minutely_15") val minutelyFifteen: OpenMeteoWeatherMinutely? = null
+data class OpenMeteoWeatherMinutely(
+    val time: LongArray,
+    //@SerialName("precipitation_probability") val precipitationProbability: Array<Int?>?,
+    val precipitation: Array<Float?>?
 )